src/main/java/com/raydya/SortIntegersByTheNumberOf1Bits.java
package com.raydya;

import java.util.ArrayList;
import java.util.List;

public class SortIntegersByTheNumberOf1Bits {

    public int[] sortByBits(int[] arr) {
        int[] bit = new int[10001];
        List<Integer> list = new ArrayList<>();
        for (int x : arr) {
            list.add(x);
            bit[x] = Integer.bitCount(x);
        }
        list.sort((x, y) -> {
            if (bit[x] != bit[y]) {
                return bit[x] - bit[y];
            } else {
                return x - y;
            }
        });
        for (int i = 0; i < arr.length; ++i) {
            arr[i] = list.get(i);
        }
        return arr;
    }
}
